create a module fore the node id stuff and the keys
